Move code from radial_bins_plots.py to the new pages adaptive_phenotyping.py and preprocessing.py
